% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_botanique}
\alias{demo_botanique}
\title{sortie_freq_botanique}
\format{
A data frame with 4392 rows and 4610 variables
}
\source{
GallicaEnv
}
\usage{
demo_botanique
}
\description{
Jeu de donnees bibliographiques.
}
\keyword{datasets}
