% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_eaux_foret}
\alias{demo_eaux_foret}
\title{sortie_freq_eaux_forets}
\format{
A data frame with 1336 rows and 2447 variables
}
\source{
GallicaEnv
}
\usage{
demo_eaux_foret
}
\description{
Jeu de donnees bibliographiques.
}
\keyword{datasets}
